Add edge_tracker and bitset_t in its own .h and .cc file (separated out from merge.*) Add kVehicularAccess to graphconstants.h.
