+++ valhalla/baldr/graphconstants.h
@@ -24,6 +24,10 @@ constexpr uint16_t kHOVAccess        = 128;
 constexpr uint16_t kWheelchairAccess = 256;
 constexpr uint16_t kAllAccess        = 4095;
 
+// Constant representing vehicular access types
+constexpr uint32_t kVehicularAccess = kAutoAccess | kTruckAccess |
+                                      kTaxiAccess | kBusAccess | kHOVAccess;
+
 // Maximum number of transit records per tile and other max. transit
 // field values.
 constexpr uint32_t kMaxTransitDepartures    = 16777215;
